refactor client POST
basically turned all the <button> elements into <input type=button> that are all part of the same form. This totally removed the need to send an AJAX POST, and reallows the ability to render output in a new view. need to finish refactoring by getting the clicked button to send is identity as part of the POST body